Slider: Add support for 'wraparound' ranges - min can be greater than the max.
The new 'wraparound' option to the slider allows you to make a slider that is
also has the range option set to 'true' able to drag the second handle before
the first. This is handy if you wanted to use the slider to select part of
something that wraps around, like the days of the week. With wraparound on, you
can select Friday-Monday as well as Monday-Friday. When the wraparound is
triggered, the slider bar classes invert, so the shaded area is between the ends
of the slider and the nearest handle, instead of between the handles.